+++ Source/VolumeCorrector.Core/Strategies/MediumCorrectionStrategy.cs
@@ -2,15 +2,15 @@
 {
     public class MediumCorrectionStrategy : ICorrectionStrategy
     {
-        public float GetTargetVolume(float volume, float loudness, float maxVolume, float maxLoudness)
+        public double GetTargetVolume(double volume, float loudness, double maxVolume, float maxLoudness)
         {
             var distance = maxLoudness - loudness;
             if (distance < 0)
                 distance = 0;
 
-            var targetAbsolute = (0.5f + distance);
-            if (targetAbsolute > 1f)
-                targetAbsolute = 1f;
+            var targetAbsolute = (0.5D + distance);
+            if (targetAbsolute > 1D)
+                targetAbsolute = 1D;
 
             return targetAbsolute * maxVolume;
         }
